requirements.txt
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
ply==3.11
python-dotenv==1.0.1
requests==2.31.0
setuptools==69.5.1
six==1.16.0
stone==3.3.3
urllib3==2.2.1
Werkzeug==3.0.3
psycopg2-binary
Flask-SQLAlchemy
Flask-Migrate
Flask-Bcrypt
Flask-JWT-Extended
authlib
